configuração de arquivos ts e ajustes eslint/pretier
