+++ todo-list/package.json
@@ -9,7 +9,7 @@
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-scripts": "5.0.0",
-    "tsc-init": "^2.1.0",
+    "tsc": "^2.0.4",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -49,6 +49,7 @@
     "eslint-plugin-promise": "^5.2.0",
     "eslint-plugin-react": "^7.29.4",
     "prettier": "^2.6.1",
+    "tsc-init": "^2.1.0",
     "typescript": "^4.6.3"
   }
 }
