+++ todo-list/src/App.tsx
@@ -1,0 +1,9 @@
+function App() {
+  return (
+    <div className="App">
+      <h1>Hello World</h1>
+    </div>
+  );
+}
+
+export default App;
