+++ todo-list/src/components/Content.tsx
@@ -1,0 +1,5 @@
+export function Content() {
+  return <></>;
+}
+
+export default Content;
