+++ todo-list/src/components/Header.tsx
@@ -1,0 +1,5 @@
+export function Header() {
+  return <></>;
+}
+
+export default Header;
